requirements.txt
asgiref==3.8.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
dj-database-url==2.1.0
Django==5.0.4
django-environ==0.11.2
django-widget-tweaks==1.5.0
gcloud==0.18.3
googleapis-common-protos==1.63.0
gunicorn==20.1.0
httplib2==0.22.0
idna==3.7
jwcrypto==1.5.6
oauth2client==4.1.3
pillow==10.3.0
protobuf==4.25.3
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pycryptodome==3.20.0
pyparsing==3.1.2
Pyrebase4==4.7.1
python-jwt==4.1.0
requests==2.29.0
requests-toolbelt==0.10.1
rsa==4.9
setuptools==69.5.1
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==1.26.18
whitenoise==6.6.0
